+++ books.php
@@ -3,7 +3,7 @@
  * Plugin Name:  Books
  * Plugin URI:   https://github.com/webmandesign/books
  * Description:  Registers new "Books" custom post type and related taxonomies.
- * Version:      1.0.0
+ * Version:      1.1.0
  * Author:       WebMan Design, Oliver Juhas
  * Author URI:   https://www.webmandesign.eu/
  * License:      GNU General Public License v3