+++ changelog.md
@@ -1,5 +1,18 @@
 # Books Changelog
 
+## 1.1.0, 20230301
+
+### Added
+- Rating taxonomy
+
+### File updates
+	books.php
+	changelog.md
+	languages/*.*
+	includes/tax/_index.php
+	includes/tax/rating.php
+
+
 ## 1.0.0, 20230110
 
 - Initial release.